desktop-client/requirements.txt
socketio-client
pyautogui